cleaned up dead functions and move file system functions to fs_utils
